+++ backend/django_react_starter/settings/base.py
@@ -147,7 +147,7 @@ DJANGO_SUPERUSER_PASSWORD = os.getenv("DJANGO_SUPERUSER_PASSWORD")
 # > Internationalization
 # --------------------------------------------------------------------------------
 # https://docs.djangoproject.com/en/4.0/topics/i18n/
-LANGUAGE_CODE = "fr-fr"
+LANGUAGE_CODE = "en-us"
 TIME_ZONE = "UTC"
 USE_I18N = True
 USE_TZ = True